pyproject.toml
[project]
name = "scoutbook-openapi-mcp"
version = "1.0.0"
description = "BSA Scoutbook MCP server auto-generated from OpenAPI spec"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=3.0.0",
    "httpx>=0.27.0",
    "python-dateutil>=2.8.0",
    "cryptography>=42.0.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
